separated handshake logic from inbound handler.
